lalala

+++ app/Http/Controllers/Admin/AdminController.php
@@ -9,8 +9,10 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\Models\Admin;
+use App\Models\Product;
 
 class AdminController extends Controller
 {
@@ -28,8 +30,33 @@ class AdminController extends Controller
        return redirect('admin/home');
     }
 
-    public function destroy($id){
-        $admin = Admin::find($id);
-        $admin->delete();
+    public function products(){
+        if(Auth::user()->role=='admin' || Auth::user()->role=='manager'){
+            $products =Product::paginate(15);
+            return view('admin.products', ['products' => $products]);
+        }
+        return redirect('admin/home');
     }
+
+    public function destroy($id){
+        if(Auth::user()->role=='admin'){
+            $admin = Admin::find($id);
+            $admin->delete();
+        }
+        else return redirect('admin/home');
+    }
+
+    public function updateManager(Request $request){
+        if(Auth::user()->role=='admin'){
+            if($request->ajax())
+            {
+                $admin = Admin::find($request->id);
+                $column=$request->name;
+                $admin->$column = $request->new_val;
+                $admin->save();
+            }
+        }
+        else return redirect('admin/home');
+    }
+
 }
